Spotitube/src/main/java/nl/han/ica/oose/dea/niels/resourcelayer/TrackResource.java
package nl.han.ica.oose.dea.niels.resourcelayer;

import nl.han.ica.oose.dea.niels.servicelayer.interfaces.ITrackService;

import javax.inject.Inject;
import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.Produces;
import javax.ws.rs.QueryParam;
import javax.ws.rs.core.MediaType;
import javax.ws.rs.core.Response;

@Path("/tracks")
public class TrackResource {
    private ITrackService trackService;

    @Inject
    public void setTrackService(ITrackService trackService) {
        this.trackService = trackService;
    }

    @GET
    @Produces(MediaType.APPLICATION_JSON)
    public Response getAllTracks(@QueryParam("token") String token, @QueryParam("forPlaylist") int forPlaylistId) {
        return Response.status(200).entity(trackService.getAllTracksForPlaylist(forPlaylistId, token)).build();
    }
}
